# tailsamplingprocessor: Optimize tag mutator memory allocations 