@jiekun 2023-10-25
> A test is failing on a place that is related to the changes:
> 
> ```
> --- FAIL: TestTraceIntegrity (0.00s)
> panic: runtime error: index out of range [0] with length 0 [recovered]
> 	panic: runtime error: index out of range [0] with length 0
> 
> goroutine 15 [running]:
> testing.tRunner.func1.2({0x112df00, 0xc00016a5a0})
> 	/opt/hostedtoolcache/go/1.20.10/x64/src/testing/testing.go:1526 +0x372
> testing.tRunner.func1()
> 	/opt/hostedtoolcache/go/1.20.10/x64/src/testing/testing.go:1529 +0x650
> panic({0x112df00, 0xc00016a5a0})
> 	/opt/hostedtoolcache/go/1.20.10/x64/src/runtime/panic.go:890 +0x263
> github.com/open-telemetry/opentelemetry-collector-contrib/processor/tailsamplingprocessor.(*tailSamplingSpanProcessor).makeDecision(0xc000326180, {0x1, 0x2, 0x3, 0x4, 0x0, 0x0, 0x0, 0x0, 0x0, ...}, ...)
> 	/home/runner/work/opentelemetry-collector-contrib/opentelemetry-collector-contrib/processor/tailsamplingprocessor/processor.go:305 +0x14b9
> github.com/open-telemetry/opentelemetry-collector-contrib/processor/tailsamplingprocessor.(*tailSamplingSpanProcessor).samplingPolicyOnTick(0xc000326180)
> ```
> 
> Anyone willing to pick this up? I can probably take a look eventually, but not _right now_.

I appreciate the optimizations made by the original author. 

The issue with the unit test only required a very small change, but since I am unable to submit it on the original branch, I have opened a separate PR https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/28597. But yes @brancz feel free to add those fix lines to the original branch and I will delete the new pr ^_^